+++ docs/CHANGELOG.rst
@@ -27,6 +27,7 @@ Added
 - Add ``get_annotations`` on ``Sample`` which returns all annotations on a
   sample as a dictionary
 - Optionally show progress bar when downloading files
+- Optionally show progress bar when uploading files
 
 Fixed
 -----